Update chat example
