+++ examples/chat/index.ts
@@ -1,0 +1,1 @@
+import './chat.js';
